move generic random types outside of indef
